+++ meerschaum/_internal/entry.py
@@ -1,6 +1,7 @@
 #! /usr/bin/env python
 # -*- coding: utf-8 -*-
 # vim:fenc=utf-8
+# type: ignore
 
 """
 The entry point for launching Meerschaum actions.
@@ -48,7 +49,7 @@ def entry(sysargs: Optional[List[str]] = None) -> SuccessTuple:
 
     if args.get('schedule', None):
         from meerschaum.utils.schedule import schedule_function
-        return schedule_function(entry_with_args, args['schedule'], **args)
+        return schedule_function(entry_with_args, **args)
     return entry_with_args(**args)
 
 
@@ -61,7 +62,7 @@ def entry_with_args(
     """
     import sys
     from meerschaum.plugins import Plugin
-    from meerschaum.actions import get_shell, get_action, get_main_action_name
+    from meerschaum.actions import get_action, get_main_action_name
     from meerschaum._internal.arguments import remove_leading_action
     from meerschaum.utils.venv import Venv, active_venvs, deactivate_venv
     if kw.get('trace', None):