+++ meerschaum/config/_default.py
@@ -44,6 +44,7 @@ default_meerschaum_config = {
             },
             'local': {
                 'host': 'localhost',
+                'port': 8000,
             },
             'mrsm': {
                 'host': 'api.mrsm.io',
@@ -151,7 +152,6 @@ default_config['pipes'] = default_pipes_config
 default_config['plugins'] = default_plugins_config
 from meerschaum.config._jobs import default_jobs_config
 default_config['jobs'] = default_jobs_config
-#  default_config['experimental'] = default_experimental_config
 ### add configs from other packages
 try:
     import meerschaum.config.stack
@@ -160,6 +160,8 @@ except ImportError as e:
 finally:
     from meerschaum.config.stack import default_stack_config
 default_config['stack'] = default_stack_config
+from meerschaum.config._dash import default_dash_config
+default_config['dash'] = default_dash_config
 
 default_header_comment = """
 #####################################################################