+++ docs/mkdocs/news/changelog.md
@@ -1,8 +1,116 @@
 # 🪵 Changelog
 
-## 2.1.x Releases
+## 2.2.x Releases
 
 This is the current release cycle, so stay tuned for future releases!
+
+### v2.2.0
+
+**New Features**
+
+- **New job scheduler**  
+  The job scheduler has been rewritten with a [simpler syntax](https://meerschaum.io/reference/background-jobs/#-schedules).
+
+  ```bash
+  mrsm sync pipes -s 'daily & mon-fri starting 00:00 tomorrow' -d
+  ```
+
+- **Add `show schedule`.**  
+  Validate your schedules' upcoming timestamps with `show schedule`.
+
+  ```
+  mrsm show schedule 'daily & mon-fri starting 2024-05-01'
+
+  Next 5 timestamps for schedule 'daily & mon-fri starting 2024-05-01':
+
+    2024-05-01 00:00:00+00:00
+    2024-05-02 00:00:00+00:00
+    2024-05-03 00:00:00+00:00
+    2024-05-06 00:00:00+00:00
+    2024-05-07 00:00:00+00:00
+  ```
+
+- **Added timestamps to log file lines.**  
+  Log files now prepend the current minute to each line of the file, and the timestamps are also printed when viewing logs with `show logs`.
+  To disable this behavio, set `MRSM{jobs:logs:timestamps:enabled}` to `false`.
+
+  You may change the timestamp format under the config keys `MRSM{jobs:logs:timestamps:format}` (timestamp written to disk) and `MRSM{jobs:logs:timestamps:follow_format}` (timestamp printed when following via `show logs`.).
+
+- **Add `--skip-deps`.**  
+  When installing plugins, you may skip dependencies with `--skip-deps`. This should improve the iteration loop during development.
+
+  ```bash
+  mrsm install plugin noaa --no-deps
+  ```
+
+- **Add logs buttons to job cards on the Web UI.**  
+  For your convenience, "Follow logs" and "Download logs" buttons have been added to jobs' cards.
+
+- **Add a Delete button to job cards on the Web UI.**  
+  You may now delete a job from its card (once stopped, that is).
+
+- **Add management buttons to pipes' cards.**  
+  For your convenience, you may now sync, verify, clear, drop, and delete pipes directly from cards.
+
+- **Designate your packages as plugins with the `meerschaum.plugins` entry point.**  
+  You may now specify your existing packages as Meerschaum plugins by adding the `meerschaum.plugins` [Entrypoint](https://packaging.python.org/en/latest/guides/creating-and-discovering-plugins/#using-package-metadata) to your package metadata:
+
+  ```python
+  from setuptools import setup
+  
+  setup(
+      ...,
+      entry_points = {
+          'meerschaum.plugins': [
+              'foo = foo',
+          ],
+      },
+  )
+  ```
+
+  or if you are using `pyproject.toml`:
+
+  ```toml
+  [project.entry-points."meerschaum.plugins"]
+  foo = "foo"
+  ```
+
+- **Pre- and post-sync hooks are printed separately.**  
+  The results of sync hooks are now printed right after execution rather than after the sync.
+
+**Bugfixes**
+
+- **Fixed a filtering bug on the Web UI when changing instances.**  
+  When changing instances on the Web Console, the connector, metric, and location choices will reset appropriately.
+
+- **Ctrl+C when exiting `show logs`.**  
+  Pressing Ctrl+C will now exit the `show logs` immediately.
+
+**Breaking Changes**
+
+- **No longer supporting the old scheduler syntax.**  
+  If you have jobs with the old scheduler syntax (e.g. using the keyword `before`), you may need to delete and recreate your jobs with an updated schedule.
+
+- **Upgraded to `psycopg` from `psycopg2`.**  
+  The upgrade to `psycopg` (version 3) should provide better performance for larger transactions.
+
+- **`Daemon.cleanup()` now returns a `SuccessTuple`.**
+
+**Other changes**
+
+- **Bumped `xterm.js` to v5.5.0.**
+- **Added tags to the pipes card.**
+- **Replaced `watchgod` with `watchfiles`.**
+- **Replaced `rocketry` with `APScheduler`.**
+- **Removed `pydantic` from dependencies.**
+- **Removed `passlib` from dependencies.**
+- **Bumped default TimescaleDB image to `latest-pg16-oss`.**
+- **Held back `duckdb` to `<0.10.3`.**
+
+
+## 2.1.x Releases
+
+The 2.1.x series added high-performance upserts, improved numerics support and temporary tables performance, and many other bugfixes and improvements.
 
 ### v2.1.7
 
@@ -32,7 +140,6 @@ This is the current release cycle, so stay tuned for future releases!
   #   color
   # 1  blue
   ```
-
 - **Add `get_in_ex_params()` to `meerschaum.utils.misc`.**  
   This function parses a standard `params` dictionary into tuples of include and exclude parameters.
 
@@ -47,19 +154,26 @@ This is the current release cycle, so stay tuned for future releases!
 
 - **Add `coerce_numeric` to `pipe.enforce_dtypes()`.**  
   Setting this to `False` will not cast floats to `Decimal` if the corresponding dtype is `int`.
+
 - **Improve JSON serialization when filtering for updates.**
+
 - **Add `date_bound_only` to `pipe.filter_existing()`.**  
   The argument `date_bound_only` means that samples retrieved by `pipe.get_data()` will only use `begin` and `end` for bounding. This may improve performance for custom instance connectors which have limited searchability.
+
 - **Add `safe_copy` to `pipe.enforce_types()`, `pipe.filter_existing()`, `filter_unseen_df()`.**  
   By default, these functions will create copies of dataframes to avoid mutating the input dataframes. Setting `safe_copy` to `False` may be more memory efficient.
+
 - **Add multiline support to `extract_stats_from_message`.**  
   Multiple messages separated by newlines may be parsed at once.
+
   ```python
   from meerschaum.utils.formatting import extract_stats_from_message
   extract_stats_from_message("Inserted 10, upserted 3\ninserted 11, upserted 4")
   # {'inserted': 21, 'updated': 0, 'upserted': 7}
   ```
+
 - **Remove `order by` check in SQL queries.**
+
 - **Improve shell startup performance by removing support for `cmd2`.**  
   The package `cmd2` never behaved properly, so support has been removed and only the built-in `cmd` powers the shell. As such, the configuration key `shell:cmd` has been removed.